+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/component/generate-components-code.ts
@@ -1,6 +1,6 @@
 import type { Components } from '../../../../types/index.js'
 import type { Config } from '../../../../config/index.js'
-import { generateZodToOpenAPISchemaDefinition } from '../../../zod/openapi/generate-zod-to-openapi-schema-definition.js'
+import { generateZodToOpenAPISchemaDefinition } from '../../../zod-to-openapi/defining/generate-zod-to-openapi-schema-definition.js'
 import { generateZod } from '../../../zod/generate-zod.js'
 import { resolveSchemasDependencies } from '../../../../core/schema/references/resolve-schemas-dependencies.js'
 import { generateZodToOpenAPI } from '../../../zod-to-openapi/index.js'
